Animations Added
Animations for Idle, Jumping (Start and in air), and walking added.

Code for landing, falling, and death animations added, we just need to drop the animations in if/when we are ready.

Landing and death sound effects added. I hope to work with alex to implement the walking sound effects